chime: libicuuc.so is replaced with libicuucts.so
The libicuuc.so blob serves for handling Unicode text and supporting internationalization.
libicuucts.so is a version of the International Components for Unicode (ICU) library that includes the Collation Tailoring Rules, which are used for sorting and comparing text. It is an extension of libicuuc.so and serves a specific purpose within the ICU library.